Actualización

+++ src/main/java/org/JavaCar/Roda.java
@@ -2,9 +2,9 @@ package org.JavaCar;
 
 public class Roda {
     private String marca;
-    private int diametre;
+    private double diametre;
 
-    public Roda(String marca, int diametre) {
+    public Roda(String marca, double diametre) {
         this.marca = marca;
         this.diametre = diametre;
     }
